third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-legacy-expected.txt

CONSOLE WARNING: 'HTMLVideoElement.webkitEnterFullScreen()' is deprecated. Please use 'Element.requestFullscreen()' instead.
HTMLVideoElement.webkitEnterFullScreen() in an iframe requires the allowfullscreen attribute

EVENT(webkitfullscreenerror)
END OF TEST
